fixed and reorganized fixed-asset-base benchmarks
