--- benchmarks/api/fabric/base/create-asset-batch.yaml
@@ -1,183 +1,0 @@
-test:
-  name: create-asset-batch
-  description: >-
-    This is a duration based benchmark targeting a Hyperledger Fabric network
-    using the `fixed-asset` NodeJS chaincode
-    contract that is interacted with via a Fabric-SDK-Node Gateway. Each test
-    round invokes the `createAssetsFromBatch` method, with successive rounds
-    increasing the batch size of the assets being added into the world state
-    database.
-  workers:
-    type: local
-    number: 4
-  rounds:
-    - label: create-asset-batch-20-8000-fixed-tps
-      description: >-
-        Test a submitTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `createAssetsFromBatch`, which
-        inserts a batch of 20 assets of size 8k bytes into the World State
-        database at a fixed TPS.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-rate
-        opts:
-          tps: 5
-      workload:
-        module: benchmarks/api/fabric/lib/batch-create-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          byteSize: 8000
-          batchSize: 20
-    - label: create-asset-batch-1-8000
-      description: >-
-        Test a submitTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `createAssetsFromBatch`, which
-        inserts a batch of 1 assets of size 8k bytes into the World State
-        database.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-load
-        opts:
-          transactionLoad: 5
-      workload:
-        module: benchmarks/api/fabric/lib/batch-create-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          byteSize: 8000
-          batchSize: 1
-    - label: create-asset-batch-10-8000
-      description: >-
-        Test a submitTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `createAssetsFromBatch`, which
-        inserts a batch of 10 assets of size 8k bytes into the World State
-        database.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-load
-        opts:
-          transactionLoad: 5
-      workload:
-        module: benchmarks/api/fabric/lib/batch-create-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          byteSize: 8000
-          batchSize: 10
-    - label: create-asset-batch-20-8000
-      description: >-
-        Test a submitTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `createAssetsFromBatch`, which
-        inserts a batch of 20 assets of size 8k bytes into the World State
-        database.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-load
-        opts:
-          transactionLoad: 5
-      workload:
-        module: benchmarks/api/fabric/lib/batch-create-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          byteSize: 8000
-          batchSize: 20
-    - label: create-asset-batch-30-8000
-      description: >-
-        Test a submitTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `createAssetsFromBatch`, which
-        inserts a batch of 30 assets of size 8k bytes into the World State
-        database.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-load
-        opts:
-          transactionLoad: 5
-      workload:
-        module: benchmarks/api/fabric/lib/batch-create-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          byteSize: 8000
-          batchSize: 30
-    - label: create-asset-batch-40-8000
-      description: >-
-        Test a submitTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `createAssetsFromBatch`, which
-        inserts a batch of 40 assets of size 8k bytes into the World State
-        database.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-load
-        opts:
-          transactionLoad: 5
-      workload:
-        module: benchmarks/api/fabric/lib/batch-create-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          byteSize: 8000
-          batchSize: 40
-    - label: create-asset-batch-50-8000
-      description: >-
-        Test a submitTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `createAssetsFromBatch`, which
-        inserts a batch of 50 assets of size 8k bytes into the World State
-        database.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-load
-        opts:
-          transactionLoad: 5
-      workload:
-        module: benchmarks/api/fabric/lib/batch-create-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          byteSize: 8000
-          batchSize: 50
-monitors:
-    resource:
-    - module: prometheus
-      options:
-        url: "http://localhost:9090"
-        metrics:
-          include: [dev-.*, couch, peer, orderer]
-          queries:
-          - name: Avg Memory (MB)
-            query: 'sum(container_memory_rss{name=~".+"}) by (name)'
-            step: 10
-            label: name
-            statistic: avg
-            multiplier: 0.000001
-          - name:   CPU (%)
-            query: sum(rate(container_cpu_usage_seconds_total{name=~".+"}[1m])) by (name)
-            step: 10
-            label: name
-            statistic: avg
-            multiplier: 100
-          - name: Network In (MB)
-            query: sum(rate(container_network_receive_bytes_total{name=~".+"}[1m])) by (name)
-            step: 10
-            label: name
-            statistic: sum
-            multiplier: 0.000001
-          - name: Network Out (MB)
-            query: sum(rate(container_network_transmit_bytes_total{name=~".+"}[1m])) by (name)
-            step: 10
-            label: name
-            statistic: sum
-            multiplier: 0.000001
-          - name: Disc Write (MB)
-            query: sum(rate(container_fs_writes_bytes_total{name=~".+"}[1m])) by (name)
-            step: 10
-            label: name
-            statistic: sum
-            multiplier: 0.000001
-          - name: Disc Read (MB)
-            query: sum(rate(container_fs_reads_bytes_total{name=~".+"}[1m])) by (name)
-            step: 10
-            label: name
-            statistic: sum
-            multiplier: 0.000001
