--- benchmarks/api/fabric/base/get-asset-batch.yaml
@@ -1,204 +1,0 @@
-test:
-  name: get-asset-batch
-  description: >-
-    This is a duration based benchmark targeting a Hyperledger Fabric network
-    using the `fixed-asset` NodeJS chaincode
-    contract that is interacted with via a Fabric-SDK-Node Gateway. Each test
-    round invokes the `getAssetsFromBatch` API method. Successive rounds create
-    and retrieve assets of larger byteSize.
-  workers:
-    type: local
-    number: 10
-  rounds:
-    - label: get-asset-batch-evaluate-1-8000
-      description: >-
-        Test an evaluateTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `getAssetsFromBatch`. This
-        method performs a getState on a batch of 1 UUID that matches an asset of
-        size 8000 bytes.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-load
-        opts:
-          transactionLoad: 20
-      workload:
-        module: benchmarks/api/fabric/lib/batch-get-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          create_sizes:
-            - 8000
-          assets: 8000
-          byteSize: 8000
-          batchSize: 1
-          consensus: false
-    - label: get-asset-batch-evaluate-10-8000
-      description: >-
-        Test an evaluateTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `getAssetsFromBatch`. This
-        method performs a getState on a batch of 10 UUIDs that each match an
-        asset of size 8000 bytes.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-load
-        opts:
-          transactionLoad: 20
-      workload:
-        module: benchmarks/api/fabric/lib/batch-get-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          noSetup: true
-          assets: 8000
-          byteSize: 8000
-          batchSize: 10
-          consensus: false
-    - label: get-asset-batch-evaluate-20-8000
-      description: >-
-        Test an evaluateTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `getAssetsFromBatch`. This
-        method performs a getState on a batch of 20 UUIDs that each match an
-        asset of size 8000 bytes.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-load
-        opts:
-          transactionLoad: 20
-      workload:
-        module: benchmarks/api/fabric/lib/batch-get-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          noSetup: true
-          assets: 8000
-          byteSize: 8000
-          batchSize: 20
-          consensus: false
-    - label: get-asset-batch-evaluate-30-8000
-      description: >-
-        Test an evaluateTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `getAssetsFromBatch`. This
-        method performs a getState on a batch of 30 UUIDs that each match an
-        asset of size 8000 bytes.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-load
-        opts:
-          transactionLoad: 20
-      workload:
-        module: benchmarks/api/fabric/lib/batch-get-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          noSetup: true
-          assets: 8000
-          byteSize: 8000
-          batchSize: 30
-          consensus: false
-    - label: get-asset-batch-evaluate-40-8000
-      description: >-
-        Test an evaluateTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `getAssetsFromBatch`. This
-        method performs a getState on a batch of 40 UUIDs that each match an
-        asset of size 8000 bytes.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-load
-        opts:
-          transactionLoad: 20
-      workload:
-        module: benchmarks/api/fabric/lib/batch-get-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          noSetup: true
-          assets: 8000
-          byteSize: 8000
-          batchSize: 40
-          consensus: false
-    - label: get-asset-batch-evaluate-50-8000
-      description: >-
-        Test an evaluateTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `getAssetsFromBatch`. This
-        method performs a getState on a batch of 50 UUIDs that each match an
-        asset of size 8000 bytes.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-load
-        opts:
-          transactionLoad: 20
-      workload:
-        module: benchmarks/api/fabric/lib/batch-get-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          noSetup: true
-          assets: 8000
-          byteSize: 8000
-          batchSize: 50
-          consensus: false
-    - label: get-asset-batch-evaluate-20-8000-fixed-tps
-      description: >-
-        Test an evaluateTransaction() Gateway method against the NodeJS
-        `fixed-asset` Smart Contract method named `getAssetsFromBatch`. This
-        method performs a getState on a batch of 20 UUIDs that each match an
-        asset of size 8000 bytes at a fixed TPS.
-      chaincodeID: fixed-asset-base
-      txDuration: 300
-      rateControl:
-        type: fixed-rate
-        opts:
-          tps: 30
-      workload:
-        module: benchmarks/api/fabric/lib/batch-get-asset.js
-        arguments:
-          chaincodeID: fixed-asset-base
-          noSetup: true
-          assets: 8000
-          byteSize: 8000
-          batchSize: 20
-          consensus: false
-monitors:
-    resource:
-    - module: prometheus
-      options:
-        url: "http://localhost:9090"
-        metrics:
-          include: [dev-.*, couch, peer, orderer]
-          queries:
-          - name: Avg Memory (MB)
-            query: 'sum(container_memory_rss{name=~".+"}) by (name)'
-            step: 10
-            label: name
-            statistic: avg
-            multiplier: 0.000001
-          - name:   CPU (%)
-            query: sum(rate(container_cpu_usage_seconds_total{name=~".+"}[1m])) by (name)
-            step: 10
-            label: name
-            statistic: avg
-            multiplier: 100
-          - name: Network In (MB)
-            query: sum(rate(container_network_receive_bytes_total{name=~".+"}[1m])) by (name)
-            step: 10
-            label: name
-            statistic: sum
-            multiplier: 0.000001
-          - name: Network Out (MB)
-            query: sum(rate(container_network_transmit_bytes_total{name=~".+"}[1m])) by (name)
-            step: 10
-            label: name
-            statistic: sum
-            multiplier: 0.000001
-          - name: Disc Write (MB)
-            query: sum(rate(container_fs_writes_bytes_total{name=~".+"}[1m])) by (name)
-            step: 10
-            label: name
-            statistic: sum
-            multiplier: 0.000001
-          - name: Disc Read (MB)
-            query: sum(rate(container_fs_reads_bytes_total{name=~".+"}[1m])) by (name)
-            step: 10
-            label: name
-            statistic: sum
-            multiplier: 0.000001
