+++ benchmarks/api/fabric/fixed-tps-runs.yaml
@@ -10,16 +10,16 @@ test:
         Test a submitTransaction() Gateway method against the NodeJS
         `fixed-asset` Smart Contract method named `createAsset`, which inserts
         an asset of size 8000 bytes into the World State database.
-      chaincodeID: fixed-asset-base
+      chaincodeID: fixed-asset
       txDuration: 300
       rateControl:
         type: fixed-rate
         opts:
           tps: 100
       workload:
-        module: benchmarks/api/fabric/lib/create-asset.js
+        module: benchmarks/api/fabric/workloads/create-asset.js
         arguments:
-          chaincodeID: fixed-asset-base
+          chaincodeID: fixed-asset
           byteSize: 8000
     - label: create-asset-batch-20-8000
       description: >-
@@ -27,16 +27,16 @@ test:
         `fixed-asset` Smart Contract method named `createAssetsFromBatch`, which
         inserts a batch of 20 assets of size 8k bytes into the World State
         database.
-      chaincodeID: fixed-asset-base
+      chaincodeID: fixed-asset
       txDuration: 300
       rateControl:
         type: fixed-rate
         opts:
           tps: 100
       workload:
-        module: benchmarks/api/fabric/lib/batch-create-asset.js
+        module: benchmarks/api/fabric/workloads/batch-create-asset.js
         arguments:
-          chaincodeID: fixed-asset-base
+          chaincodeID: fixed-asset
           byteSize: 8000
           batchSize: 20
     - label: get-asset-evaluate-8000
@@ -44,16 +44,16 @@ test:
         Test an evaluateTransaction() Gateway method against the NodeJS
         `fixed-asset` Smart Contract method named `getAsset`. This method
         performs a getState on an item that matches an asset of size 8000 bytes.
-      chaincodeID: fixed-asset-base
+      chaincodeID: fixed-asset
       txDuration: 300
       rateControl:
         type: fixed-rate
         opts:
           tps: 100
       workload:
-        module: benchmarks/api/fabric/lib/get-asset.js
+        module: benchmarks/api/fabric/workloads/get-asset.js
         arguments:
-          chaincodeID: fixed-asset-base
+          chaincodeID: fixed-asset
           create_sizes:
             - 100
             - 1000
@@ -72,16 +72,16 @@ test:
         `fixed-asset` Smart Contract method named `getAssetsFromBatch`. This
         method performs a getState on a batch of 20 UUIDs that each match an
         asset of size 8000 bytes.
-      chaincodeID: fixed-asset-base
+      chaincodeID: fixed-asset
       txDuration: 300
       rateControl:
         type: fixed-rate
         opts:
           tps: 50
       workload:
-        module: benchmarks/api/fabric/lib/batch-get-asset.js
+        module: benchmarks/api/fabric/workloads/batch-get-asset.js
         arguments:
-          chaincodeID: fixed-asset-base
+          chaincodeID: fixed-asset
           create_sizes:
             - 8000
           assets: 8000
@@ -94,16 +94,16 @@ test:
         `fixed-asset` Smart Contract method named `paginatedRangeQuery`. This
         method performs a paginated range query, with a passed pagesize of 20
         and a range keys that bound 200 assets created by the calling client.
-      chaincodeID: fixed-asset-base
+      chaincodeID: fixed-asset
       txDuration: 300
       rateControl:
         type: fixed-rate
         opts:
           tps: 30
       workload:
-        module: benchmarks/api/fabric/lib/mixed-range-query-asset.js
+        module: benchmarks/api/fabric/workloads/mixed-range-query-asset.js
         arguments:
-          chaincodeID: fixed-asset-base
+          chaincodeID: fixed-asset
           range: 200
           offset: 100
           pagesize: '20'
