+++ benchmarks/api/fabric/get-asset.yaml
@@ -22,7 +22,7 @@ test:
         opts:
           transactionLoad: 50
       workload:
-        module: benchmarks/api/fabric/lib/get-asset.js
+        module: benchmarks/api/fabric/workloads/get-asset.js
         arguments:
           chaincodeID: fixed-asset
           create_sizes:
@@ -49,7 +49,7 @@ test:
         opts:
           transactionLoad: 50
       workload:
-        module: benchmarks/api/fabric/lib/get-asset.js
+        module: benchmarks/api/fabric/workloads/get-asset.js
         arguments:
           chaincodeID: fixed-asset
           noSetup: true
@@ -68,7 +68,7 @@ test:
         opts:
           transactionLoad: 50
       workload:
-        module: benchmarks/api/fabric/lib/get-asset.js
+        module: benchmarks/api/fabric/workloads/get-asset.js
         arguments:
           chaincodeID: fixed-asset
           noSetup: true
@@ -87,7 +87,7 @@ test:
         opts:
           transactionLoad: 50
       workload:
-        module: benchmarks/api/fabric/lib/get-asset.js
+        module: benchmarks/api/fabric/workloads/get-asset.js
         arguments:
           chaincodeID: fixed-asset
           noSetup: true
@@ -106,7 +106,7 @@ test:
         opts:
           transactionLoad: 50
       workload:
-        module: benchmarks/api/fabric/lib/get-asset.js
+        module: benchmarks/api/fabric/workloads/get-asset.js
         arguments:
           chaincodeID: fixed-asset
           noSetup: true
@@ -126,7 +126,7 @@ test:
         opts:
           transactionLoad: 50
       workload:
-        module: benchmarks/api/fabric/lib/get-asset.js
+        module: benchmarks/api/fabric/workloads/get-asset.js
         arguments:
           chaincodeID: fixed-asset
           noSetup: true
@@ -146,7 +146,7 @@ test:
         opts:
           transactionLoad: 50
       workload:
-        module: benchmarks/api/fabric/lib/get-asset.js
+        module: benchmarks/api/fabric/workloads/get-asset.js
         arguments:
           chaincodeID: fixed-asset
           noSetup: true
@@ -166,7 +166,7 @@ test:
         opts:
           transactionLoad: 50
       workload:
-        module: benchmarks/api/fabric/lib/get-asset.js
+        module: benchmarks/api/fabric/workloads/get-asset.js
         arguments:
           chaincodeID: fixed-asset
           noSetup: true
@@ -186,7 +186,7 @@ test:
         opts:
           tps: 350
       workload:
-        module: benchmarks/api/fabric/lib/get-asset.js
+        module: benchmarks/api/fabric/workloads/get-asset.js
         arguments:
           chaincodeID: fixed-asset
           noSetup: true
