+++ benchmarks/api/fabric/mixed-range-query-pagination.yaml
@@ -17,16 +17,16 @@ test:
         `fixed-asset` Smart Contract method named `paginatedRangeQuery`. This
         method performs a paginated range query, with a passed pagesize of 10
         and a range keys that bound 200 assets created by the calling client.
-      chaincodeID: fixed-asset-base
+      chaincodeID: fixed-asset
       txDuration: 300
       rateControl:
         type: fixed-load
         opts:
           transactionLoad: 2
       workload:
-        module: benchmarks/api/fabric/lib/mixed-range-query-asset.js
+        module: benchmarks/api/fabric/workloads/mixed-range-query-asset.js
         arguments:
-          chaincodeID: fixed-asset-base
+          chaincodeID: fixed-asset
           create_sizes:
             - 100
             - 1000
@@ -48,16 +48,16 @@ test:
         `fixed-asset` Smart Contract method named `paginatedRangeQuery`. This
         method performs a paginated range query, with a passed pagesize of 20
         and a range keys that bound 200 assets created by the calling client.
-      chaincodeID: fixed-asset-base
+      chaincodeID: fixed-asset
       txDuration: 300
       rateControl:
         type: fixed-load
         opts:
           transactionLoad: 2
       workload:
-        module: benchmarks/api/fabric/lib/mixed-range-query-asset.js
+        module: benchmarks/api/fabric/workloads/mixed-range-query-asset.js
         arguments:
-          chaincodeID: fixed-asset-base
+          chaincodeID: fixed-asset
           range: 200
           offset: 100
           pagesize: '20'
@@ -69,16 +69,16 @@ test:
         `fixed-asset` Smart Contract method named `paginatedRangeQuery`. This
         method performs a paginated range query, with a passed pagesize of 50
         and a range keys that bound 200 assets created by the calling client.
-      chaincodeID: fixed-asset-base
+      chaincodeID: fixed-asset
       txDuration: 300
       rateControl:
         type: fixed-load
         opts:
           transactionLoad: 2
       workload:
-        module: benchmarks/api/fabric/lib/mixed-range-query-asset.js
+        module: benchmarks/api/fabric/workloads/mixed-range-query-asset.js
         arguments:
-          chaincodeID: fixed-asset-base
+          chaincodeID: fixed-asset
           range: 200
           offset: 100
           pagesize: '50'
@@ -90,16 +90,16 @@ test:
         `fixed-asset` Smart Contract method named `paginatedRangeQuery`. This
         method performs a paginated range query, with a passed pagesize of 100
         and a range keys that bound 200 assets created by the calling client.
-      chaincodeID: fixed-asset-base
+      chaincodeID: fixed-asset
       txDuration: 300
       rateControl:
         type: fixed-load
         opts:
           transactionLoad: 2
       workload:
-        module: benchmarks/api/fabric/lib/mixed-range-query-asset.js
+        module: benchmarks/api/fabric/workloads/mixed-range-query-asset.js
         arguments:
-          chaincodeID: fixed-asset-base
+          chaincodeID: fixed-asset
           range: 200
           offset: 100
           pagesize: '100'
@@ -111,16 +111,16 @@ test:
         `fixed-asset` Smart Contract method named `paginatedRangeQuery`. This
         method performs a paginated range query, with a passed pagesize of 200
         and a range keys that bound 200 assets created by the calling client.
-      chaincodeID: fixed-asset-base
+      chaincodeID: fixed-asset
       txDuration: 300
       rateControl:
         type: fixed-load
         opts:
           transactionLoad: 2
       workload:
-        module: benchmarks/api/fabric/lib/mixed-range-query-asset.js
+        module: benchmarks/api/fabric/workloads/mixed-range-query-asset.js
         arguments:
-          chaincodeID: fixed-asset-base
+          chaincodeID: fixed-asset
           range: 200
           offset: 100
           pagesize: '200'
@@ -132,16 +132,16 @@ test:
         `fixed-asset` Smart Contract method named `paginatedRangeQuery`. This
         method performs a paginated range query, with a passed pagesize of 500
         and a range keys that bound 200 assets created by the calling client.
-      chaincodeID: fixed-asset-base
+      chaincodeID: fixed-asset
       txDuration: 300
       rateControl:
         type: fixed-load
         opts:
           transactionLoad: 2
       workload:
-        module: benchmarks/api/fabric/lib/mixed-range-query-asset.js
+        module: benchmarks/api/fabric/workloads/mixed-range-query-asset.js
         arguments:
-          chaincodeID: fixed-asset-base
+          chaincodeID: fixed-asset
           range: 200
           offset: 100
           pagesize: '300'
@@ -154,16 +154,16 @@ test:
         method performs a paginated range query, with a passed pagesize of 20
         and a range keys that bound 200 assets created by the calling client at
         a fixed TPS.
-      chaincodeID: fixed-asset-base
+      chaincodeID: fixed-asset
       txDuration: 300
       rateControl:
         type: fixed-rate
         opts:
           tps: 10
       workload:
-        module: benchmarks/api/fabric/lib/mixed-range-query-asset.js
+        module: benchmarks/api/fabric/workloads/mixed-range-query-asset.js
         arguments:
-          chaincodeID: fixed-asset-base
+          chaincodeID: fixed-asset
           range: 200
           offset: 100
           pagesize: '20'
