+++ benchmarks/api/fabric/workloads/batch-create-asset.js
