+++ benchmarks/api/fabric/workloads/batch-delete-asset.js
