+++ benchmarks/api/fabric/workloads/batch-get-asset.js
