+++ benchmarks/api/fabric/workloads/create-asset.js
