+++ benchmarks/api/fabric/workloads/delete-asset.js
