+++ benchmarks/api/fabric/workloads/helper.js
