+++ benchmarks/api/fabric/workloads/mixed-range-query-asset.js
