+++ benchmarks/api/fabric/workloads/mixed-rich-query-asset.js
