+++ benchmarks/api/fabric/workloads/range-query-asset.js
