+++ benchmarks/api/fabric/workloads/rich-query-asset.js
